Update mantainer, description and dependencies
The `debian/control` file has been updated with the correct name and
email to contact the mantainer. Changes in description and the inclusion
of `debconf` as a dependency were made attempting to make the package
Lintian clean (issue #4).

The control file has also been updated with VCS information, indicating
details about the main Git repository.

- update mantainer's email and name
- fixes `missing-debconf-dependency` (Lintian)
- fixes `synopsis-is-a-sentence` (Lintian)
- adds repository details (Vcs-Browser and Vcs-Git)